grammar: rename Offset to OffsetField to avoid collision with Node.Offset
Note, this issue was especially subtle as the Textmapper
reported no error when generating the parser and the
new Offset method had the same signature as the
Node.Offset, thus the compiler reported no error. However,
the returned offset of OffsetField would of course be
unrelated to the offset in the source file, as expected
by the Node interface.

Note to self, change back from OffsetField to Offset once
inspirer/textmapper#13 is resolved.

Updates inspirer/textmapper#13.