grammar: re-work handling of align in global and function declarations
This is not a proper fix, but works for the majority of cases.

If anyone has a better idea how to solve the LR-1 conflicts
related to Align in FuncAttribute and as part of GlobalDecl
and FuncDef, FuncDecl, please do get in touch.

Fixes llir/llvm#40. Well, most of it at least.
We still need to figure out how to handle Align,
AttrString and AttrPair of ReturnAttribute.